widgetise attachment conversations component
