+++ likeminds-chat-reactnative-integration/ChatSX/components/MessageNotSupported/index.tsx
@@ -1,0 +1,16 @@
+import { View, Text } from "react-native";
+import React from "react";
+import { styles } from "../AttachmentConversations/styles";
+import { MESSAGE_NOT_SUPPORTED } from "../../constants/Strings";
+
+const MessageNotSupported = () => {
+  return (
+    <View>
+      <Text style={styles.deletedMsg}>
+        {MESSAGE_NOT_SUPPORTED}
+      </Text>
+    </View>
+  );
+};
+
+export default MessageNotSupported;
